Closes trac:2824.
Expands app_idx to int32_t for large numbers of app_contexts.

This commit was SVN r24941.

The following Trac tickets were found above:
  Ticket 2824 --> https://svn.open-mpi.org/trac/ompi/ticket/2824

+++ orte/runtime/data_type_support/orte_dt_print_fns.c
@@ -527,8 +527,8 @@ int orte_dt_print_app_context(char **output, char *prefix, orte_app_context_t *s
         asprintf(&pfx2, "%s", prefix);
     }
     
-    asprintf(&tmp, "\n%sData for app_context: index %lu\tapp: %s\n%s\tNum procs: %lu",
-             pfx2, (unsigned long)src->idx, src->app,
+    asprintf(&tmp, "\n%sData for app_context: index %d\tapp: %s\n%s\tNum procs: %lu",
+             pfx2, src->idx, src->app,
              pfx2, (unsigned long)src->num_procs);
     
     count = opal_argv_count(src->argv);
